refactor(CnosDB, CockroachDB, Materialize, Postgres): remove duplicates from DBAlias
