+++ src/sqlancer/cnosdb/ast/CnosDBAlias.java
@@ -1,35 +1,18 @@
 package sqlancer.cnosdb.ast;
 
-import sqlancer.common.visitor.UnaryOperation;
+import sqlancer.common.ast.DBAlias;
 
-public class CnosDBAlias implements UnaryOperation<CnosDBExpression>, CnosDBExpression {
+public class CnosDBAlias extends DBAlias<CnosDBExpression> implements CnosDBExpression {
 
     private final CnosDBExpression expr;
-    private final String alias;
 
     public CnosDBAlias(CnosDBExpression expr, String alias) {
+        super(alias);
         this.expr = expr;
-        this.alias = alias;
     }
 
     @Override
     public CnosDBExpression getExpression() {
         return expr;
     }
-
-    @Override
-    public String getOperatorRepresentation() {
-        return " as " + alias;
-    }
-
-    @Override
-    public OperatorKind getOperatorKind() {
-        return OperatorKind.POSTFIX;
-    }
-
-    @Override
-    public boolean omitBracketsWhenPrinting() {
-        return true;
-    }
-
 }
