+++ src/sqlancer/cockroachdb/ast/CockroachDBAlias.java
@@ -1,35 +1,18 @@
 package sqlancer.cockroachdb.ast;
 
-import sqlancer.common.visitor.UnaryOperation;
+import sqlancer.common.ast.DBAlias;
 
-public class CockroachDBAlias implements UnaryOperation<CockroachDBExpression>, CockroachDBExpression {
+public class CockroachDBAlias extends DBAlias<CockroachDBExpression> implements CockroachDBExpression {
 
     private final CockroachDBExpression expr;
-    private final String alias;
 
     public CockroachDBAlias(CockroachDBExpression expr, String alias) {
+        super(alias);
         this.expr = expr;
-        this.alias = alias;
     }
 
     @Override
     public CockroachDBExpression getExpression() {
         return expr;
     }
-
-    @Override
-    public String getOperatorRepresentation() {
-        return " as " + alias;
-    }
-
-    @Override
-    public OperatorKind getOperatorKind() {
-        return OperatorKind.POSTFIX;
-    }
-
-    @Override
-    public boolean omitBracketsWhenPrinting() {
-        return true;
-    }
-
 }
