+++ src/sqlancer/materialize/ast/MaterializeAlias.java
@@ -1,35 +1,18 @@
 package sqlancer.materialize.ast;
 
-import sqlancer.common.visitor.UnaryOperation;
+import sqlancer.common.ast.DBAlias;
 
-public class MaterializeAlias implements UnaryOperation<MaterializeExpression>, MaterializeExpression {
+public class MaterializeAlias extends DBAlias<MaterializeExpression> implements MaterializeExpression {
 
     private final MaterializeExpression expr;
-    private final String alias;
 
     public MaterializeAlias(MaterializeExpression expr, String alias) {
+        super(alias);
         this.expr = expr;
-        this.alias = alias;
     }
 
     @Override
     public MaterializeExpression getExpression() {
         return expr;
     }
-
-    @Override
-    public String getOperatorRepresentation() {
-        return " as " + alias;
-    }
-
-    @Override
-    public OperatorKind getOperatorKind() {
-        return OperatorKind.POSTFIX;
-    }
-
-    @Override
-    public boolean omitBracketsWhenPrinting() {
-        return true;
-    }
-
 }
