+++ src/sqlancer/postgres/ast/PostgresAlias.java
@@ -1,35 +1,18 @@
 package sqlancer.postgres.ast;
 
-import sqlancer.common.visitor.UnaryOperation;
+import sqlancer.common.ast.DBAlias;
 
-public class PostgresAlias implements UnaryOperation<PostgresExpression>, PostgresExpression {
+public class PostgresAlias extends DBAlias<PostgresExpression> implements PostgresExpression {
 
     private final PostgresExpression expr;
-    private final String alias;
 
     public PostgresAlias(PostgresExpression expr, String alias) {
+        super(alias);
         this.expr = expr;
-        this.alias = alias;
     }
 
     @Override
     public PostgresExpression getExpression() {
         return expr;
     }
-
-    @Override
-    public String getOperatorRepresentation() {
-        return " as " + alias;
-    }
-
-    @Override
-    public OperatorKind getOperatorKind() {
-        return OperatorKind.POSTFIX;
-    }
-
-    @Override
-    public boolean omitBracketsWhenPrinting() {
-        return true;
-    }
-
 }
